Display pixels in an even grid
Uses a calculation to determine flex-basis for each pixel and
sets that number as an inline style on each pixel. There may be a
way to do this that is "nicer" than setting an inline style.
Using CSS Grid for this would be better, but at this point Odin
Project hasn't had the learner use Grid, so solving problems such
as this (with the "current" tools available) has become a learning
process/puzzle on it's own.